get unit tests to pass on 32-bit systems

+++ src/Phing/Type/Selector/DateSelector.php
@@ -99,7 +99,7 @@ class DateSelector extends BaseExtendSelector
      */
     public function setMillis($millis)
     {
-        $this->setSeconds((int) $millis / 1000);
+        $this->setSeconds((float) $millis / 1000);
     }
 
     /**
